Code climat
* Readme update
* Code climat connect
* Binary Heap Refactor
