+++ lib/Heap.rb
@@ -1,6 +1,7 @@
 require 'Heap/version'
 
 module Heap
+  require 'Heap/binary_heap/binary_heap'
   require 'Heap/binary_heap/binary_heap_min'
   require 'Heap/binary_heap/binary_heap_max'
   require 'Heap/multiple_heap/multiple_heap_max'
