+++ lib/Heap/binary_heap/binary_heap_max.rb
@@ -1,31 +1,7 @@
 module Heap
   module BinaryHeap
-    class MaxHeap
-      attr_reader :elements
-
-      def initialize(elements = [])
-        @elements = []
-        add(elements.pop) until elements.empty?
-      end
-
-      def add(element)
-        if element.is_a? Array
-          element.each do |el|
-            @elements.push el
-            swim_up(count)
-          end
-        elsif defined? element.elements
-          add element.elements
-        else
-          @elements.push element
-          swim_up(count)
-        end
-      end
-
-      def count
-        @elements.length
-      end
-
+    # Binary Heap with max root
+    class MaxHeap < BinaryHeap
       def extract_max
         @elements[0]
       end
@@ -47,12 +23,6 @@ module Heap
 
       private
 
-      def swap(index1, index2)
-        temp = @elements[index1 - 1]
-        @elements[index1 - 1] = @elements[index2 - 1]
-        @elements[index2 - 1] = temp
-      end
-
       def swim_up(index)
         return if index == 1
         parent_index = index / 2
@@ -62,32 +32,12 @@ module Heap
       end
 
       def swim_down(index)
-        child1_index = 2 * index
-        child2_index = 2 * index + 1
-        return if @elements[child1_index - 1].nil? && @elements[child2_index - 1].nil?
-        if @elements[child2_index - 1].nil?
-          return if @elements[child1_index - 1] <= @elements[index - 1]
-          swap(index, child1_index)
-          swim_down(child1_index)
-        else
-          if @elements[child2_index - 1] <= @elements[index - 1] && @elements[child1_index - 1] <= @elements[index - 1]
-            return
-          elsif @elements[child2_index - 1] > @elements[index - 1] && @elements[child1_index - 1] > @elements[index - 1]
-            if @elements[child2_index - 1] > @elements[child1_index - 1]
-              swap(child2_index, index)
-              swim_down(child2_index)
-            else
-              swap(child1_index, index)
-              swim_down(child1_index)
-            end
-          elsif @elements[child2_index - 1] > @elements[index - 1]
-            swap(child2_index, index)
-            swim_down(child2_index)
-          elsif @elements[child1_index - 1] > @elements[index - 1]
-            swap(child1_index, index)
-            swim_down(child1_index)
-          end
-        end
+        children = get_children(index)
+        return if children.empty?
+        max_child = children.max
+        return if @elements[index - 1] >= max_child[0]
+        swap index, max_child[1]
+        swim_down max_child[1]
       end
     end
   end
